+++ cmd/werf/chart/ts/init/init.go
@@ -99,8 +99,9 @@ func runChartTSInit(ctx context.Context, chartDir string) error {
 	})
 
 	if err := action.ChartTSInit(ctx, action.ChartTSInitOptions{
-		ChartDirPath: chartPath,
-		ChartName:    werfConfig.Meta.Project,
+		ChartDirPath:      chartPath,
+		ChartName:         werfConfig.Meta.Project,
+		RenderContextType: nelmcommon.TSWerfRenderContextType,
 	}); err != nil {
 		return fmt.Errorf("chart ts init: %w", err)
 	}