+++ cmd/werf/common/container_backend.go
@@ -136,11 +136,6 @@ func InitProcessContainerBackend(ctx context.Context, cmdData *CmdData, registry
 			return nil, ctx, fmt.Errorf("unable to get buildah client: %w", err)
 		}
 
-		err = docker.InitDockerConfig(docker.InitOptions{DockerConfigDir: *cmdData.DockerConfig})
-		if err != nil {
-			return nil, ctx, fmt.Errorf("unable to set docker config for buildah client: %w", err)
-		}
-
 		return wrapContainerBackend(container_backend.NewBuildahBackend(b, container_backend.BuildahBackendOptions{TmpDir: filepath.Join(werf.GetServiceDir(), "tmp", "buildah")})), ctx, nil
 	}